Basic_News_Website_wFlask/requirements.txt
asn1crypto==1.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
jsonify==0.5
MarkupSafe==1.1.1
newsapi-python==0.2.6
pycparser==2.19
pyOpenSSL==19.1.0
PySocks==1.7.1
requests==2.22.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
win-inet-pton==1.1.0
wincertstore==0.2
